+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for Python 3.14 (#99).
 - Generic oscilloscope driver.
 - Type annotations in emulator code.
 - Emulator module tests.
+- Display measure function to Keithley 2657A emulator.
 
 ### Changed
 
@@ -20,9 +22,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set default emulator host to `localhost`.
 - Renamed this changelog to `CHANGELOG.md`.
 
+### Removed
+
+- Obsolete `mypy.ini` (not required with recent mypy versions).
+
 ### Fixed
 
 - Emulator routes are now properly overriden from derived classes in function `get_routes`.
+
+## [1.3.1] - 2025-10-23
+
+### Fixed
+
+- PILAS emulator missing response from tune command (#100).
 
 ## [1.3.0] - 2025-08-11
 
@@ -108,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ITK CorvisTT emulator.
 
-[unreleased]: https://github.com/hephy-dd/comet/releases/tag/v1.3.0...HEAD
+[unreleased]: https://github.com/hephy-dd/comet/releases/tag/v1.3.1...HEAD
+[1.3.1]: https://github.com/hephy-dd/comet/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/hephy-dd/comet/compare/v1.2.2...v1.3.0
 [1.2.2]: https://github.com/hephy-dd/comet/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/hephy-dd/comet/compare/v1.2.0...v1.2.1